Убрал Either. Ответ парсится не в double, а в Long.

+++ app/src/main/java/ru/mavrinvladislav/shifttask2025/authorization/data/remote/dto/CreateOtpResponse.kt
@@ -8,5 +8,5 @@ data class CreateOtpResponse(
     @SerializedName("reason")
     val reason: String,
     @SerializedName("retryDelay")
-    val retryDelay: Double
+    val retryDelay: Long
 )
